chore: alterando caminho da home
